Moved validation to RouteBuilder and removed RouteParameters
