+++ UPGRADE.md
@@ -2,11 +2,11 @@
 
 ## From 0.9.* to 0.10.0
 
-### Moved validation from `RoutePaylaod` to `RouteParameters` and removed `RoutePayload::generate`
+### Moved route parameter validation to `RouteBuilder` and made it mandatory
 
-Using the `RouteBuilder` is now **highly recommended**. There is no breaking change if you're already using it. 
+Using the `RouteBuilder` is now **mandatory**. The validation has been moved to `addCommandRoute` and `addQueryRoute`. The `RouteParameters` have been removed in favor of parameters directly for the functions. 
 
-When not using it, you have to replace your usages of `RoutePayload::generate` (which has been removed) with `RoutePayload::generatePayloadFromRouteParameters`.
+When not using it yet, you have to replace your usages of `RoutePayload::generate` (which has been removed) with the `RouteBuilder` functions.
 
 Before:
 
@@ -19,50 +19,50 @@ $routes->add(
     ->methods([Request::METHOD_POST])
     ->defaults([
         'routePayload' => RoutePayload::generate(
-            ...
+            dtoClass: CreateProductNewsArticleCommand::class,
+            handlerClass: CreateProductNewsArticleCommandHandler::class,
         ),
+    ]);
 ```
 
 After:
 
 ```php
-$routes->add(
-    'api_news_create_news_article_command',
-    '/api/news/create-news-article-command',
-)
-    ->controller([CommandController::class, 'handle'])
-    ->methods([Request::METHOD_POST])
-    ->defaults([
-        'routePayload' => RoutePayload::generateFromRouteParameters(new RouteParameters(
-            ...
-        )),
+RouteBuilder::addCommandRoute(
+    $routes,
+    path: '/api/news/create-news-article-command',
+    dtoClass: CreateProductNewsArticleCommand::class,
+    handlerClass: CreateProductNewsArticleCommandHandler::class,
+);
 ```
 
 ## The route name generation changed
 
-The route name generation changed. When the name must be something specific (because it's used as a reference), it must be set as a parameter for `RouteParameters`. The name generation might change in future versions. If the name isn't used anywhere, you nether need to nor should set it.
+The route name generation changed. When the name must be something specific (because it's used as a reference), it must be set as a parameter for `addCommandRoute` and `addQueryRoute`. The name generation might change in future versions. If the name isn't used anywhere, you nether need to nor should set it.
 
-**Only if a name of a route is used as a reference**, add it as a parameter to `RouteParameters`.
+**Only if a name of a route is used as a reference**, add it as a parameter to `addCommandRoute` and `addQueryRoute`.
 
 Before:
 
 ```php
-RouteBuilder::addCommandRoute($routes, new RouteParameters(
+RouteBuilder::addCommandRoute(
+    $routes,
     path: '/api/news/create-news-article-command',
     dtoClass: CreateProductNewsArticleCommand::class,
     handlerClass: CreateProductNewsArticleCommandHandler::class,
-));
+);
 ```
 
 After:
 
 ```php
-RouteBuilder::addCommandRoute($routes, new RouteParameters(
+RouteBuilder::addCommandRoute(
+    $routes,
     path: '/api/news/create-news-article-command',
     dtoClass: CreateProductNewsArticleCommand::class,
     handlerClass: CreateProductNewsArticleCommandHandler::class,
     name: 'api_news_create_news_article_command',
-));
+);
 ```
 
 ## From 0.8.* to 0.9.0
