+++ docs/routing.md
@@ -2,40 +2,13 @@
 
 For better typing and refactoring, the routing must be configured with PHP files instead of the usual YAML files.
 
-When the routes are generated (on cache warmup), they are cached as PHP files. Therefore, the configuration can't contain any object instances. For increased type safety and better DX, we use a value object with named parameters to configure our routes. All parameters are validated on cache warmup.
+When the routes are generated (on cache warmup), they are cached as PHP files. Therefore, the configuration can't contain any object instances. For increased type safety and better DX, we use a route builder with named parameters to configure our routes. All parameters are validated on cache warmup.
 
 ```php
 return static function (RoutingConfigurator $routes) {
 
-    $routes->add(
-        'api_news_create_news_article_command',
-        '/api/news/create-news-article-command',
-    )
-        ->controller([CommandController::class, 'handle'])
-        ->methods([Request::METHOD_POST])
-        ->defaults([
-            'routePayload' => RoutePayload::generateFromRouteParameters(new RouteParameters(
-                dtoClass: CreateProductNewsArticleCommand::class,
-                handlerClass: CreateProductNewsArticleCommandHandler::class,
-                requestDecoderClass: CommandWithFilesRequestDecoder::class,
-                dtoValidatorClasses: [
-                    UserIdValidator::class => null,
-                ],
-            )),
-        ]);
-    ...
-    
-};
-```
-
-All parameters except `dtoClass` and `handlerClass` are optional. You might only need to define those when you only need the default components for the other parameters ([configured in the `cqrs.yaml`](./configuration.md)).
-
-To reduce the noise in the routes even further, it's **highly recommended** to use the `RouteBuilder`. It chooses the controller depending on the function used, uses `POST` as default method, generates the name from the path and handles the route payload generation for you. The above example can be rewritten like this:
-
-```php
-return static function (RoutingConfigurator $routes) {
-
-    RouteBuilder::addCommandRoute($routes, new RouteParameters(
+    RouteBuilder::addCommandRoute(
+        $routes,
         path: '/api/news/create-news-article-command',
         dtoClass: CreateProductNewsArticleCommand::class,
         handlerClass: CreateProductNewsArticleCommandHandler::class,
@@ -43,11 +16,15 @@ return static function (RoutingConfigurator $routes) {
         dtoValidatorClasses: [
             UserIdValidator::class => null,
         ],
-    ));
+    );
     ...
     
 };
 ```
+
+All parameters except `path`, `dtoClass` and `handlerClass` are optional. You might only need to define those when you only need the default components for the other parameters ([configured in the `cqrs.yaml`](./configuration.md)).
+
+The `RouteBuilder` chooses the controller depending on the function used (`addCommandRoute` or `addQueryRoute`), uses `POST` as default method, generates the name based on the path and validates all parameters. The routes are constructed on cache warmup, so that's the only time the validation costs performance.
 
 ## Overwrite request validators, request data transformers, DTO validators and handler wrappers
 
@@ -66,7 +43,7 @@ return static function (CqrsConfig $cqrsConfig) {
 And you then adapt it like this:
 
 ```php
-RouteBuilder::addCommandRoute($routes, new RouteParameters(
+RouteBuilder::addCommandRoute(
     ...
     dtoValidatorClasses: [
         FilesizeValidator::class => null,
@@ -79,7 +56,7 @@ The end result is that only the `FilesizeValidator` is left. You need to include
 This way you're also able to remove all default DTO validators from a route like this:
 
 ```php
-RouteBuilder::addCommandRoute($routes, new RouteParameters(
+RouteBuilder::addCommandRoute(
     ...
     dtoValidatorClasses: [],
 ));
@@ -102,7 +79,7 @@ return static function (CqrsConfig $cqrsConfig) {
 And you then use the following route configuration:
 
 ```php
-RouteBuilder::addCommandRoute($routes, new RouteParameters(
+RouteBuilder::addCommandRoute(
     ...
     dtoValidatorClassesToMergeWithDefault: [
         CourseIdValidator::class => null,
@@ -130,7 +107,7 @@ return static function (CqrsConfig $cqrsConfig) {
 And the following route configuration:
 
 ```php
-RouteBuilder::addCommandRoute($routes, new RouteParameters(
+RouteBuilder::addCommandRoute(
     ...
     dtoValidatorClassesToMergeWithDefault: [
         FilesizeValidator::class => 10,
